requirements_preprocess.txt
torch>=2.5.0
torchaudio>=2.5.0
numpy>=1.24.0
onnxruntime-gpu>=1.16.0
pypinyin>=0.49.0
python-dotenv>=1.0.0
jieba>=0.42.1
runpod>=1.1.0